requirements.txt
pokemontcgsdk
SQLAlchemy
Flask
python-dotenv